requirements.txt
importlib-metadata>=1.7.0
numpy>=1.18.5
psutil>=5.7.0
torch
torchvision
future>=0.18.2
importlib-metadata>=1.7.0
joblib>=0.15.1
numpy>=1.18.5
Pillow>=8.3.2
psutil>=5.7.0
pydantic>=1.8.2
PyYAML>=5.3.1
redis>=3.5.3
typing>=3.7.4.1
httptools>=0.1.1
python-json-logger>=2.0.1
loguru>=0.5.3
requests>=2.25.1
grpcio>=1.32.0
protobuf